requirements.txt
google-generativeai
linkedin-api
requests
Flask
Pillow
